Formatter improvements
