+++ shards/tools/shlangbox/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+clap = { version = "4.4.7", features = ["derive"] }
 pest = "2.7.1"
 shards-lang-core = { path = "../../lang_core" }
